+++ omod/src/test/java/org/openmrs/module/radiology/order/web/RadiologyOrderFormControllerTest.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology.web.controller;
+package org.openmrs.module.radiology.order.web;
 
 import static org.hamcrest.collection.IsMapContaining.hasKey;
 import static org.hamcrest.core.Is.is;
